Additions and fixes
Fixes inability to connect to some Gemini capsules like midnight.pub.  IPv6 support added.  Fix leak of SSL pointer.  Headings are now wrapped in a nicer way; having a hanging indent until the first space; this makes e.g. numbered/sectioned headings to look very nice.  Directory listings added via file:// protocol, they simply generate a gemtext buffer.  Bandaid fix for memory overflow when rendering pages that render more than can fit in the buffer, we simply create a larger initial buffer.  The size is a multiple of the size that was received.  This fix should work in almost all cases; there is theoretically some edge cases where if enough indenting is somehow caused then it could possibly overflow it, but it is unlikely.  Doing a proper reallocation would be a huge pain because the line buffers, etc. store loads of pointers to the buffer and a realloc would throw all the pointers to free'd memory.  May need to find a solution to this.  'e' keybind in normal mode was fixed and now also edits the selected link if there is one. URI parsing code may have been made a little more robust.  Also a fix for a subtle edge case for pages that have a single line with no LF at the end (last char was not rendered before).