implements various things and some fixes
MIME types added; text/plain and text/gemini are only supported for now (may add support for mailcap for viewing non-renderable content).  MIME type is shown on right-side of status line.  Lot's of code cleanup, especially in typesetter and some of the 'request' code.  Adds support for 'parent' directory and 'root' directory via keybinds.  Some work on path normalisation for URIs (doesn't work perfectly just yet but well enough to not cause too many issues).  'file://' protocol code moved to a new local.c file (in lieu of a better name).  New interesting but subtle feature added to gemtext renderer where links which are part of a 'group' and have an index with less digits are aligned to the one with the most digits.  It looks alright though in future may be altered to align to largest in the file rather than in 'groups'.  Escape codes for headings are also no longer lost over wrapped lines when the initial escape is out of the visible buffer.  Cross-protocol links are now indicated in a way similar to AV-98.  util code moved to an actual source file rather than being header-only