on-disk caching!
On-disk caching has now been implemented and works fairly well.  The system is maybe a little clunky; a single 'meta.dir' file is stored to list all files that are cached in a file tree, and lists metadata about them.  It seems that there are some memory bugs that need to be tracked down though, as the 'meta.dir' sometimes gets garbage text in it.  The 'meta.dir' seems to also sometimes get duplicate entries for some reason?  Selection and scrolling is now restored properly and stored with cached pages in memory rather than in history.  Some scrolling bugs fixed.  Since cached pages (at least when built with on-disk support) now store checksums, an indicator is shown on refresh to indicate if the resource has changed.  In the TUI input prompt, it is now possible to use Ctrl+W to backspace larger portions of text at a time.  Ctrl+P keybind was added too for input prompt to cycle a detected protocol between 'gemini://' and 'gopher://' (may need some work though).