support for INPUT and SENSITIVE status codes
also implements URI query parsing, (and now also ignores fragments if found).  URIs with querys are not cached at the moment.  Queries are encoded with proper RFC 3986 compliance.  SENSITIVE input prompts are slightly different in that they echo asterisks rather than the actual chars.  Typesetter doesn't re-typeset on resize unless specifically the width changes.  Searching functionality extended; though 'n' and 'p' keys for link selection are in a slightly awkward state, we need to find a better choice of keybinds for it.  Searching has been decided to operate on the typeset buffer only rather than the raw one; to avoid serious complications, though we still have to deal with one annoyance, that is the searching of text which has been hyphenated by the typesetter; or text which has had extra spaces included in the line.  These issues will need to be addressed.  Search match 'invalidation' also needs a proper implementation; at the moment searching is broken on window resizes.  Page reload functionality re-implemented.