New justification algorithm! +more
justification is now performed in a way that seems similar to how nroff justifies; we 'alternate' between distributing the space between left and right of the lines.  This produces a very nice distribution and often looks a lot better than the other (often more complicated) methods that were played with.  Indentation code for gemtext has been refactored a lot and justification now works for list items, where it was previously broken for the first line.  Cached item age is now shown in the bottom-right when loading from cache.  The message is brief but works well enough for now.