utf8 line length fixes, finally
utf8_strnlen_w_formats now replaced with utf8_width function, which uses wcswidth to find the number of terminal cells the string occupies.  This obviously makes the function a bit slower; especially as there's an extra step of stripping escapes and tabs, but nevertheless it works and pages now render properly.  One annoyance has been realised, that is that some Gemtext documents which use a series of hyphens, equal signs, etc. that are not preformatted, for some reason do not get rendered properly.