Initial support for 'favourites' file and easy interaction with it
at the moment input for favourite title is not supported; however such titles are read from the file as whatever comes after the URI itself.  Two modes of favouriting are supported; adding via 'toggle' and via a 'prompt'.  Deletion of favourites can also take place on the favourites page when viewed in-buffer (is loaded on startup) by pressing 'D' while a link on the page is selected.